Move locale translation loading into the bundles.
- Move responsibility for loading locale-specific translations into the
  bundles themselves. This obviates the need for both the `bundlePath`
  property and `@dojo/core/load`.
- Update `cldr/load` to accept only CLDR data objects directly.
- Remove `cldr/load/webpack`.
- Update the README and doc comments.
