+++ src/main.ts
@@ -13,10 +13,13 @@ import {
 import i18n, {
 	Bundle,
 	formatMessage,
+	FormatOptions,
 	getCachedMessages,
 	getMessageFormatter,
 	I18n,
 	invalidate,
+	LocaleLoaders,
+	LocaleTranslations,
 	MessageFormatter,
 	Messages,
 	setLocaleMessages,
@@ -66,6 +69,7 @@ export {
 	formatDate,
 	formatMessage,
 	formatNumber,
+	FormatOptions,
 	formatRelativeTime,
 	formatUnit,
 	generateLocales,
@@ -82,6 +86,8 @@ export {
 	I18n,
 	invalidate,
 	loadCldrData,
+	LocaleLoaders,
+	LocaleTranslations,
 	MessageFormatter,
 	Messages,
 	normalizeLocale,
