--- tests/support/mocks/common/es/main.ts
@@ -1,5 +1,0 @@
-export = {
-	hello: 'Hola',
-	helloReply: 'Hola',
-	goodbye: 'Adiós'
-};
