+++ tests/unit/main.ts
@@ -24,7 +24,6 @@ import i18n, {
 	parseDate,
 	parseNumber,
 	pluralize,
-	setLocaleMessages,
 	switchLocale,
 	systemLocale
 } from '../../src/main';
@@ -55,7 +54,6 @@ registerSuite('main', {
 		assert.isFunction(parseDate, 'parseDate is exported.');
 		assert.isFunction(parseNumber, 'parseNumber is exported.');
 		assert.isFunction(pluralize, 'pluralize is exported.');
-		assert.isFunction(setLocaleMessages, 'setLocaleMessages is exported.');
 		assert.isFunction(switchLocale, 'switchLocale is exported.');
 		assert.isString(systemLocale, 'systemLocale is exported.');
 	}
